tests/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0 